Leveraged existing delegated Stack class with StackWrapper class.
