+++ Makefile
@@ -1,10 +1,10 @@
 CXXFLAGS = -std=c++11
 
 all: main.o
-	g++ -std=c++11 -o a.out main.o
+	g++ -std=c++11 -o a.out wrapper_main.o
 
-main.o: main.cpp Stack_Node.h StackWrapper.h
-	g++ -std=c++11 -c main.cpp
+main.o: wrapper_main.cpp Stack.h StackWrapper.h
+	g++ -std=c++11 -c wrapper_main.cpp
 
 
 clean:
