+++ StackWrapper.h
@@ -1,7 +1,7 @@
 #ifndef STACKWRAP_H
 #define STACKWRAP_H
 
-#include"Stack_Node.h"
+#include"Stack.h"
 
 
 /*
