+++ wrapper_main.cpp
@@ -12,7 +12,6 @@
 using namespace std;
 
 #include"StackWrapper.h"
-#include"Stack_Node.h"
